requirements.txt
python-dotenv~=1.0.0
langchain~=0.0.174
pydantic~=1.10.7